pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cortex-memory"
version = "0.1.0"
description = "Hybrid cognitive memory system for AI agents — episodic + semantic."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["ai", "agents", "memory", "rag", "llm", "embeddings", "obsidian"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Core
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.12",
    # Vector DB
    "chromadb>=0.5",
    # Local embeddings (default backend)
    "sentence-transformers>=3.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.25"]
ollama = ["httpx>=0.27"]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "ruff",
    "mypy",
    "pre-commit",
]
all = ["cortex-memory[openai,anthropic,ollama,dev]"]

[project.scripts]
cortex = "cortex.cli.main:app"

[tool.setuptools.package-data]
"cortex" = ["agent_guidelines.md", "skills/**/*.md"]

[project.urls]
Homepage = "https://github.com/yourusername/cortex"
Repository = "https://github.com/yourusername/cortex"
Issues = "https://github.com/yourusername/cortex/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["cortex*"]

# ── Ruff ──────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

# ── Mypy ──────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true

# ── Pytest ────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=cortex --cov-report=term-missing -q"
